+++ src/events/emitter.js
@@ -1,3 +1,0 @@
-import LocalEmitter from "./LocalEmitter";
-
-export default new LocalEmitter();